fix grammar and avoid join game without players, update db and add gameVersion to known how electronic version is being used
