+++ lib/config/router/guards/game_guard.dart
@@ -1,6 +1,6 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:monopoly_banker/config/router/monopoly_router.gr.dart';
-import 'package:monopoly_banker/data/core/monopoly_electronico/monopoly_electronico_bloc.dart';
+import 'package:monopoly_banker/data/core/monopoly_electronico/banker_electronic_bloc.dart';
 import 'package:monopoly_banker/data/service/secure_storage.dart';
 import 'package:monopoly_banker/data/service_locator.dart';
 
@@ -15,7 +15,7 @@ class GameGuard extends AutoRouteGuard {
     } else {
       final id = await getIt<MonopolyGamesStorage>().idSesion();
       if (id != null) {
-        getIt<MonopolyElectronicoBloc>().add(RestoreGameEvent(sesionId: id));
+        getIt<MonopolyElectronicBloc>().add(RestoreGameEvent(sessionId: id));
         resolver.redirect(const ElectronicGameRoute());
         return;
       }
