+++ lib/data/service/secure_storage.dart
@@ -3,14 +3,14 @@ import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 class MonopolyGamesStorage {
 // Create storage
   final storage = const FlutterSecureStorage();
-  static const String _eletronicGames = 'saved_games_x';
-  static const String _sesionId = 'id_sesion';
+  static const String _electronicGames = 'saved_games_x';
+  static const String _sessionId = 'id_session';
   static const String _classicGames = 'saved_games_c';
 
   /// Has games saved [bool]
   Future<bool> get hasCurrentGames async {
     final futures = [
-      storage.read(key: _eletronicGames),
+      storage.read(key: _electronicGames),
       storage.read(key: _classicGames),
     ];
 
@@ -19,9 +19,9 @@ class MonopolyGamesStorage {
     return responses.contains('1');
   }
 
-  Future<void> startGameX({required String sesionId}) async {
-    await storage.write(key: _sesionId, value: sesionId);
-    await storage.write(key: _eletronicGames, value: '1');
+  Future<void> startGameX({required String sessionId}) async {
+    await storage.write(key: _sessionId, value: sessionId);
+    await storage.write(key: _electronicGames, value: '1');
   }
 
   Future<void> saveGameClassic() async {
@@ -29,12 +29,12 @@ class MonopolyGamesStorage {
   }
 
   Future<String?> idSesion() async {
-    return await storage.read(key: _sesionId);
+    return await storage.read(key: _sessionId);
   }
 
   Future<void> deleteGameX() async {
-    await storage.delete(key: _sesionId);
-    await storage.delete(key: _eletronicGames);
+    await storage.delete(key: _sessionId);
+    await storage.delete(key: _electronicGames);
   }
 
   Future<void> deleteGameClassic() async {
