+++ lib/data/service_locator.dart
@@ -1,7 +1,7 @@
 import 'package:get_it/get_it.dart';
 import 'package:monopoly_banker/config/router/monopoly_router.dart';
-import 'package:monopoly_banker/data/core/monopoly_electronico/monopoly_electronico_bloc.dart';
-import 'package:monopoly_banker/data/service/monopoly_electronic_service.dart';
+import 'package:monopoly_banker/data/core/monopoly_electronico/banker_electronic_bloc.dart';
+import 'package:monopoly_banker/data/service/banker_electronic_service.dart';
 import 'package:monopoly_banker/data/service/secure_storage.dart';
 import 'package:nfc_manager/nfc_manager.dart';
 
@@ -10,9 +10,9 @@ NfcManager _manager = NfcManager.instance;
 
 Future<void> servicelocator() async {
   getIt.registerSingleton(_manager);
-  final monservice = await MonopolyElectronicService.initDbX();
+  final monservice = await BankerElectronicService.initDbX();
   getIt.registerSingleton(monservice);
   getIt.registerSingleton(RouterCubit());
   getIt.registerSingleton(MonopolyGamesStorage());
-  getIt.registerSingleton(MonopolyElectronicoBloc());
+  getIt.registerSingleton(MonopolyElectronicBloc());
 }
