+++ lib/interface/views/eletronic_game/end_game.dart
@@ -2,7 +2,7 @@ import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:monopoly_banker/config/router/monopoly_router.dart';
 import 'package:monopoly_banker/data/model/monopoly_player.dart';
-import 'package:monopoly_banker/data/service/monopoly_electronic_service.dart';
+import 'package:monopoly_banker/data/service/banker_electronic_service.dart';
 import 'package:monopoly_banker/data/service/secure_storage.dart';
 import 'package:monopoly_banker/data/service_locator.dart';
 import 'package:monopoly_banker/interface/widgets/monopoly_credit_card.dart';
@@ -24,7 +24,7 @@ class _EndGameMonopolyXState extends State<EndGameMonopolyX> {
   }
 
   void goHome() {
-    getIt<MonopolyElectronicService>().resetPlayers();
+    getIt<BankerElectronicService>().resetPlayers();
     getIt<MonopolyGamesStorage>().deleteGameX();
     getIt<RouterCubit>().goHome();
   }
