+++ lib/interface/views/eletronic_game/game_screen.dart
@@ -1,0 +1,40 @@
+import 'package:auto_route/annotations.dart';
+import 'package:flutter/material.dart';
+import 'package:monopoly_banker/config/utils/game_versions_support.dart';
+import 'package:monopoly_banker/interface/views/eletronic_game/setup_game_electronic.dart';
+
+@RoutePage()
+class GameScreen extends StatelessWidget {
+  const GameScreen({super.key, required this.version});
+  final GameVersions version;
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: getScreen(version),
+    );
+  }
+
+  Widget getScreen(GameVersions version) {
+    switch (version) {
+      case GameVersions.classic:
+        return const ClassicGameSetup();
+      case GameVersions.electronic:
+        return const ElectronicGameSetup();
+      case GameVersions.colima:
+        // TODO: Handle this case.
+        throw Exception("This game doesn't exist jet");
+    }
+  }
+}
+
+class ClassicGameSetup extends StatelessWidget {
+  const ClassicGameSetup({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(),
+      body: const Center(child: Text('Cooming soon')),
+    );
+  }
+}
